Added initExportSettings method.
