+++ lib/chart.js
@@ -26,6 +26,7 @@ import {
   toBoolean,
   wrapAround
 } from './utils.js';
+import { initExportSettings } from './config.js';
 
 let allowCodeExecution = false;
 
@@ -314,13 +315,16 @@ const exportAsString = (stringToExport, options, endCallback) => {
 /**
  * Starts an exporting process
  *
- * @param {object} options - All options object.
+ * @param {object} settings - Settings for export.
  * @param {function} endCallback - The function to call when exporting is done.
  */
 export default {
-  startExport: async (options, endCallback) => {
+  startExport: async (settings, endCallback) => {
     // Starting exporting process message
     log(4, '[chart] Starting exporting process.');
+
+    // Initialize options 
+    const options = initExportSettings(settings);
 
     // Get the export options
     const exportOptions = options.export;
