+++ lib/config.js
@@ -19,7 +19,7 @@ import prompts from 'prompts';
 import { log } from './logger.js';
 import { mergeConfigOptions } from './utils.js';
 
-import { promptsConfig } from './schemas/config.js';
+import { defaultConfig, promptsConfig } from './schemas/config.js';
 
 /**
  * Loads the configuration from JSON file.
@@ -183,8 +183,38 @@ export const initDefaultOptions = (items) => {
   return options;
 };
 
+/**
+ * Initializes options for the `startExport` method by merging user options
+ * with the default options.
+ *
+ * @param {any} userOptions - User options for exporting.
+ * @return {object} - User options merged with default options.
+ */
+export const initExportSettings = (userOptions) => {
+  let options = {};
+
+  if (userOptions.svg) {
+    options = initDefaultOptions(defaultConfig);
+    options.export.type = userOptions.type || userOptions.export.type;
+    options.export.scale = userOptions.scale || userOptions.export.scale;
+    options.payload = {
+      svg: userOptions.svg
+    };
+  } else {
+    options = mergeConfigOptions(
+      initDefaultOptions(defaultConfig),
+      userOptions,
+      ['options', 'resources', 'globalOptions', 'themeOptions']
+    );
+  }
+
+  options.export.outfile = options.export?.outfile || `chart.${options.export?.type || '.png'}`;
+  return options;
+}
+
 export default {
   loadConfigFile,
   manualConfiguration,
-  initDefaultOptions
+  initDefaultOptions,
+  initExportSettings
 };
