+++ lib/index.js
@@ -39,8 +39,10 @@ export default {
       ['options', 'globalOptions', 'themeOptions', 'resources']
     ),
   initPool: async (options = {}) => {
+    const defaultOptions = initDefaultOptions(defaultConfig);
+
     // Load an optional config file
-    options = await loadConfigFile(options);
+    options = await loadConfigFile(mergeConfigOptions(defaultOptions, options));
 
     // Set the allowCodeExecution per export module scope
     chart.setAllowCodeExecution(
